[modify]stop timer and time tick range

+++ screen_off/idle_monitor.h
@@ -22,7 +22,7 @@ signals:
     void IdleTime(int);
 
 private:
-    unsigned long GetLastIdleTime();
+    unsigned long long GetLastIdleTime();
     void OnTimer();
 
     std::atomic<unsigned long long> _lastIdleTime;
